# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mana Rythu API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  farmer: 'farmer',
  buyer: 'buyer',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  verified: boolean;
  /** @nullable */
  rating?: number | null;
  ratingCount: number;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  createdAt: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  farmer: 'farmer',
  buyer: 'buyer',
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: RegisterInputRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface VerifyUserInput {
  verified: boolean;
}

export interface RatingInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}

export type CropStatus = typeof CropStatus[keyof typeof CropStatus];


export const CropStatus = {
  growing: 'growing',
  ready: 'ready',
  harvested: 'harvested',
} as const;

export interface Crop {
  id: number;
  farmerId: number;
  /** @nullable */
  farmerName?: string | null;
  cropName: string;
  /** @nullable */
  sowDate?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  status: CropStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type CropInputStatus = typeof CropInputStatus[keyof typeof CropInputStatus];


export const CropInputStatus = {
  growing: 'growing',
  ready: 'ready',
  harvested: 'harvested',
} as const;

export interface CropInput {
  cropName: string;
  /** @nullable */
  sowDate?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  status: CropInputStatus;
  /** @nullable */
  notes?: string | null;
}

export type CropUpdateStatus = typeof CropUpdateStatus[keyof typeof CropUpdateStatus];


export const CropUpdateStatus = {
  growing: 'growing',
  ready: 'ready',
  harvested: 'harvested',
} as const;

export interface CropUpdate {
  cropName?: string;
  /** @nullable */
  sowDate?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  status?: CropUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export interface Listing {
  id: number;
  farmerId: number;
  /** @nullable */
  farmerName?: string | null;
  /** @nullable */
  farmerVerified?: boolean | null;
  /** @nullable */
  farmerRating?: number | null;
  cropName: string;
  price: number;
  quantity: number;
  unit: string;
  location: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  available: boolean;
  createdAt: string;
}

export interface ListingInput {
  cropName: string;
  price: number;
  quantity: number;
  unit: string;
  location: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ListingUpdate {
  cropName?: string;
  price?: number;
  quantity?: number;
  unit?: string;
  location?: string;
  /** @nullable */
  description?: string | null;
  available?: boolean;
}

export interface ContactInput {
  message: string;
  buyerName: string;
  /** @nullable */
  buyerPhone?: string | null;
}

export interface ContactResponse {
  success: boolean;
  /** @nullable */
  farmerPhone?: string | null;
  message: string;
}

export interface DashboardSummary {
  totalFarmers: number;
  totalBuyers: number;
  totalListings: number;
  totalCrops: number;
  activeListings: number;
  verifiedFarmers: number;
  avgFarmerRating: number;
}

export type MarketPriceTrend = typeof MarketPriceTrend[keyof typeof MarketPriceTrend];


export const MarketPriceTrend = {
  up: 'up',
  down: 'down',
  stable: 'stable',
} as const;

export interface MarketPrice {
  cropName: string;
  price: number;
  unit: string;
  trend: MarketPriceTrend;
  changePercent: number;
  location: string;
}

export type RecommendedCropDemandLevel = typeof RecommendedCropDemandLevel[keyof typeof RecommendedCropDemandLevel];


export const RecommendedCropDemandLevel = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface RecommendedCrop {
  cropName: string;
  season: string;
  demandLevel: RecommendedCropDemandLevel;
  avgPrice: number;
  unit: string;
  reason: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface WeatherSummary {
  condition: string;
  temperature: number;
  humidity: number;
  rainfall: number;
  location: string;
  advisory: string;
}

export type AdminStatsTopCropsItem = {
  cropName: string;
  count: number;
};

export interface AdminStats {
  totalUsers: number;
  pendingVerifications: number;
  totalRevenue: number;
  activeListings: number;
  newUsersThisWeek: number;
  topCrops: AdminStatsTopCropsItem[];
}

export type GetUsersParams = {
role?: string;
verified?: boolean;
};

export type GetCropsParams = {
farmerId?: number;
status?: string;
};

export type GetListingsParams = {
farmerId?: number;
cropName?: string;
location?: string;
minPrice?: number;
maxPrice?: number;
};