# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mana Rythu API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  message: string;
  buyerName: string;
  /** @nullable */
  buyerPhone?: string | null;
}