# lib/api-zod/src/generated/types/crop.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mana Rythu API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CropStatus } from './cropStatus';

export interface Crop {
  id: number;
  farmerId: number;
  /** @nullable */
  farmerName?: string | null;
  cropName: string;
  /** @nullable */
  sowDate?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  status: CropStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}