# lib/api-zod/src/generated/types/cropInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mana Rythu API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CropInputStatus } from './cropInputStatus';

export interface CropInput {
  cropName: string;
  /** @nullable */
  sowDate?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  status: CropInputStatus;
  /** @nullable */
  notes?: string | null;
}