# lib/api-zod/src/generated/types/cropUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mana Rythu API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CropUpdateStatus } from './cropUpdateStatus';

export interface CropUpdate {
  cropName?: string;
  /** @nullable */
  sowDate?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  status?: CropUpdateStatus;
  /** @nullable */
  notes?: string | null;
}